stock_control_backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-filter==25.1
django-cors-headers==4.7.0
django-environ==0.12.0
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
requests==2.31.0